Add files via upload
ComboScript 5.1 is the main file.  It imports VisioNameTool's contents in, must be in the same folder location.
The main function starts in a KeyPress while loop. Depending on the input, 1 or 2, the code runs the Visio Name Tool or the AutoColoring. 
The VisioNameTool is code straight from the ComboScript3.0, WITHOUT the Creator() function method and the Pause() method. Also removed the bottom GUI() call, will now be called in the main() while loop.
Includes 2 classes. The first class: Options has all the main code from the original AutoScripts. Added an area for functionality for saving coloring data to an excel sheet and calculating the audit totals. The second class: AutoColor houses the GUI for the actual coloring. The user chooses what the majority color of APs are and the code will color the rest.

Main() runs the whole program